pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cloudaudit"
version = "0.1.0"
description = "Cloud Security Audit Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "pytest",
    "httpx"
]

[project.scripts]
server = "server.app:main"

